chore: refactor functions from mediaPlayback to jw-media
