+++ src/components/dialog/DialogPublicTalkMediaPicker.vue
@@ -98,7 +98,7 @@ import type {
 import BaseDialog from 'components/dialog/BaseDialog.vue';
 import { storeToRefs } from 'pinia';
 import { errorCatcher } from 'src/helpers/error-catcher';
-import { unzipJwpub } from 'src/helpers/mediaPlayback';
+import { unzipJwpub } from 'src/helpers/jw-media';
 import { getPublicationsPath } from 'src/utils/fs';
 import { findDb } from 'src/utils/sqlite';
 import { useCurrentStateStore } from 'stores/current-state';
